Kp 3586

+++ planner/src/main/java/com/planner/service/vrpService/VRPGeneratorService.java
@@ -85,7 +85,7 @@ public class VRPGeneratorService {
         });
         List<Task> tasks = new ArrayList<>(vrpTaskPlanningDTO.getTasks().size());
         vrpTaskPlanningDTO.getTasks().forEach(t->{
-            tasks.add(new Task(t.getId(),t.getInstallationNumber(),t.getLatitude(),t.getLongitude(),t.getSkills(),t.getDuration(),t.getStreetName(),t.getHouseNo(),t.getBlock(),t.getFloorNo(),t.getPost(),t.getCity(),false));
+            tasks.add(new Task(t.getId(),t.getInstallationNumber(),t.getLatitude(),t.getLongitude(),t.getSkills(),(int)t.getDuration(),t.getStreetName(),t.getHouseNo(),t.getBlock(),t.getFloorNo(),t.getPost(),t.getCity(),false));
         });
         Object[] objects= getEmployeesAndShifts(vrpTaskPlanningDTO.getShifts());
         List<Shift> shifts = (List<Shift>)objects[0];
